Decompose deck_stats_panel/properties.py into mixin + chart-html + constants
properties.py conflated three independent concerns: static data tables,
a stateless HTML/CSS chart-rendering subsystem, and the panel mixin. Split
them into siblings to match how this package factors helpers:

- stats_constants.py: colour/type/hand tables + mana SVG loading.
- stats_chart_html.py: CSS/JS templates, colour math, HTML builders
  (_lerp_hex, _curve_colour, _build_vbars/_build_hbars/_build_html,
  _EMPTY_HTML).
- properties.py: now just DeckStatsPanelPropertiesMixin (673 -> 196 LOC).
- _hypergeometric_exactly moved to utils/math_utils.py as the lenient
  hypergeometric_exactly (returns 0.0 for out-of-range instead of raising,
  preserving _hand_items behaviour).

handlers.py and frame.py updated to import _EMPTY_HTML/_build_html from
stats_chart_html. Added unit tests for hypergeometric_exactly.
